+++ api/src/main/java/org/allaymc/api/player/ClientState.java
@@ -1,26 +1,26 @@
-package org.allaymc.api.network;
+package org.allaymc.api.player;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
 /**
- * Represents the status of a client.
+ * Represents the state of a client.
  *
  * @author daoge_cmd
  */
 @AllArgsConstructor
-public enum ClientStatus {
-    // The previous status of DISCONNECTED can be CONNECTED, LOGGED_IN or IN_GAME
+public enum ClientState {
+    // The previous state of DISCONNECTED can be CONNECTED, LOGGED_IN or IN_GAME
     DISCONNECTED(null),
     CONNECTED(DISCONNECTED),
     LOGGED_IN(CONNECTED),
     IN_GAME(LOGGED_IN);
 
     /**
-     * The previous status of the client. Can be {@code null} if previous status is not exist.
+     * The previous state of the client. Can be {@code null} if previous state is not exist.
      */
     @Getter
-    private final ClientStatus previousStatus;
+    private final ClientState previousState;
 
     public boolean canHandlePackets() {
         return this != DISCONNECTED;